LPS-54880 moves java
