+++ modules/apps/layout/layout-prototype-web/src/com/liferay/layout/prototype/web/portlet/LayoutPrototypePortlet.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutprototypes;
+package com.liferay.layout.prototype.web.portlet;
 
 import com.liferay.portal.NoSuchLayoutPrototypeException;
 import com.liferay.portal.RequiredLayoutPrototypeException;
