+++ modules/apps/layout/layout-prototype-web/src/com/liferay/layout/prototype/web/portlet/lar/LayoutPrototypePortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutprototypes.lar;
+package com.liferay.layout.prototype.web.portlet.lar;
 
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.lar.BasePortletDataHandler;
